requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
pymssql==2.2.10
pysftp==0.2.10
paramiko==3.3.1
schedule==1.2.1
APScheduler==3.10.4
python-dotenv==1.0.0
Werkzeug==2.3.7
pandas==2.1.3
email-validator==2.1.0
cryptography==41.0.5
Flask-APScheduler==1.13.1
pytest==7.4.3
gunicorn==21.2.0
loguru==0.7.2
